+++ gatsby-config.js
@@ -10,9 +10,9 @@ const siteURL = 'https://visiosto.netlify.app';
 module.exports = {
   siteMetadata: {
     siteURL,
+    // The wrong case of the constant must be included as some plugins depend on it.
     siteUrl: siteURL,
-    alternativeURL: 'https://anttikivi.com',
-    defaultURL: 'https://anttikivi.netlify.app',
+    alternativeURLs: ['https://anttikivi.com', 'https://anttikivi.netlify.app'],
     title: 'Antti Kivi',
     description: 'Helsinkiläinen yrittäjä',
     twitterAuthor: '@anttikiwi',