+++ src/components/App.jsx
@@ -5,7 +5,16 @@
 // https://github.com/joshwcomeau/dark-mode-minimal
 
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import ThemeContextProvider from './ThemeContextProvider';
 
-export default (props) => <ThemeContextProvider>{props.children}</ThemeContextProvider>;
+const propTypes = { children: PropTypes.any.isRequired };
+
+function App({ children }) {
+  return <ThemeContextProvider>{children}</ThemeContextProvider>;
+}
+
+App.propTypes = propTypes;
+
+export default App;