+++ src/components/Theme.jsx
@@ -2,8 +2,17 @@
 // Licensed under the MIT License
 
 import React from 'react';
+import PropTypes from 'prop-types';
 import { ThemeProvider } from 'styled-components';
 
 import theme from '../theme';
 
-export default (props) => <ThemeProvider theme={theme}>{props.children}</ThemeProvider>;
+const propTypes = { children: PropTypes.node.isRequired };
+
+function Theme({ children }) {
+  return <ThemeProvider theme={theme}>{children}</ThemeProvider>;
+}
+
+Theme.propTypes = propTypes;
+
+export default Theme;