+++ src/components/layout/Head.jsx
@@ -2,20 +2,41 @@
 // Licensed under the MIT License
 
 import React from 'react';
+import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 import { useStaticQuery, graphql } from 'gatsby';
 import { useIntl } from 'react-intl';
 
-import createIntl from '../../util/createIntl';
+import createInternationalization from '../../util/createInternationalization';
 import createLocaleURL from '../../util/createLocaleURL';
 
-const Head = (props) => {
+const propTypes = {
+  article: PropTypes.bool,
+  author: PropTypes.object,
+  description: PropTypes.string,
+  errorPage: PropTypes.bool,
+  home: PropTypes.bool,
+  image: PropTypes.object,
+  locale: PropTypes.string.isRequired,
+  pageID: PropTypes.string.isRequired,
+  title: PropTypes.string.isRequired,
+};
+
+const defaultProps = {
+  article: false,
+  author: null,
+  description: '',
+  errorPage: false,
+  home: false,
+  image: null,
+};
+
+function Head({ article, author, description, errorPage, home, image, locale, pageID, title }) {
   const data = useStaticQuery(
     graphql`
       query {
         site {
           siteMetadata {
-            defaultLocale
             description
             facebookAppID
             locales
@@ -49,81 +70,70 @@ const Head = (props) => {
 
   const { siteMetadata } = data.site;
 
-  const i = createIntl(useIntl());
+  const intl = createInternationalization(useIntl());
   const baseURL =
     process.env.NODE_ENV === 'development' ? 'http://localhost:8000' : siteMetadata.siteURL;
 
-  const titleTemplate = `%s - ${siteMetadata.title}`;
-  const title = `${props.title} - ${siteMetadata.title}`;
-  const description = props.description || i('metaDescription');
+  const titleTemplate = home
+    ? `${siteMetadata.title} - ${intl('headSlogan')}`
+    : `%s - ${siteMetadata.title}`;
+  const pageTitle = home
+    ? `${siteMetadata.title} - ${intl('headSlogan')}`
+    : `${title} - ${siteMetadata.title}`;
+  const pageDescription = description === '' ? intl('headDescription') : description;
 
   return (
     <Helmet titleTemplate={titleTemplate}>
-      <html lang={`${siteMetadata.simpleLocales[props.locale.replace('-', '_')]}`} />
-      <title>{props.title}</title>
+      <html lang={`${siteMetadata.simpleLocales[locale.replace('-', '_')]}`} />
+      <title>{title}</title>
 
-      <meta name="description" content={description} />
+      <meta name="description" content={pageDescription} />
 
       <meta property="fb:app_id" content={siteMetadata.facebookAppID} />
 
-      <meta property="og:title" content={title} />
-      <meta property="og:description" content={description} />
+      <meta property="og:title" content={pageTitle} />
+      <meta property="og:description" content={pageDescription} />
       <meta
         property="og:type"
         content={(() => {
-          if (props.article) {
+          if (article) {
             return 'article';
-          } else if (props.author) {
+          } else if (author) {
             return 'profile';
           }
           return 'website';
         })()}
       />
-      <meta
-        property="og:image"
-        content={props.image ? props.image.file.url : `${baseURL}/thumbnail.png`}
-      />
+      <meta property="og:image" content={image ? image.file.url : `${baseURL}/thumbnail.png`} />
       <meta
         property="og:image:secure_url"
-        content={props.image ? props.image.file.url : `${baseURL}/thumbnail.png`}
+        content={image ? image.file.url : `${baseURL}/thumbnail.png`}
       />
-      <meta
-        property="og:image:type"
-        content={props.image ? props.image.file.contentType : 'image/png'}
-      />
-      <meta
-        property="og:image:alt"
-        content={props.image ? props.image.description : i('metaOgImageAlt')}
-      />
+      <meta property="og:image:type" content={image ? image.file.contentType : 'image/png'} />
+      <meta property="og:image:alt" content={image ? image.description : intl('headOGImageText')} />
       <meta property="og:site_name" content={siteMetadata.title} />
       {(() => {
-        if (props.errorPage) {
+        if (errorPage) {
           const pagePath =
-            props.locale === siteMetadata.defaultLocale
+            locale === siteMetadata.defaultLocale
               ? '404'
-              : `${siteMetadata.localePaths[props.locale.replace('-', '_')]}/404`;
+              : `${siteMetadata.localePaths[locale.replace('-', '_')]}/404`;
           return <meta property="og:url" content={`${baseURL}/${pagePath}`} />;
         } else {
           return (
-            <meta
-              property="og:url"
-              content={createLocaleURL(baseURL, props.pageId, props.locale, data)}
-            />
+            <meta property="og:url" content={createLocaleURL(baseURL, pageID, locale, data)} />
           );
         }
       })()}
-      <meta
-        property="og:locale"
-        content={props.locale === 'fi' ? 'fi_FI' : props.locale.replace('-', '_')}
-      />
+      <meta property="og:locale" content={locale === 'fi' ? 'fi_FI' : locale.replace('-', '_')} />
 
       {siteMetadata.locales
-        .filter((locale) => locale !== props.locale)
-        .map((locale) => (
+        .filter((listLocale) => listLocale !== locale)
+        .map((listLocale) => (
           <meta
-            key={locale}
+            key={listLocale}
             property="og:locale:alternate"
-            content={locale === 'fi' ? 'fi_FI' : locale.replace('-', '_')}
+            content={listLocale === 'fi' ? 'fi_FI' : listLocale.replace('-', '_')}
           />
         ))}
 
@@ -131,42 +141,41 @@ const Head = (props) => {
       <meta name="twitter:site" content={siteMetadata.twitterAuthor} />
       <meta
         name="twitter:creator"
-        content={
-          props.author && props.author.twitter
-            ? `@${props.author.twitter}`
-            : siteMetadata.twitterAuthor
-        }
+        content={author && author.twitter ? `@${author.twitter}` : siteMetadata.twitterAuthor}
       />
 
       <link rel="stylesheet" href="https://use.typekit.net/fis6wxg.css" />
 
       {(() => {
-        if (props.errorPage) {
-          return siteMetadata.locales.map((locale) => (
+        if (errorPage) {
+          return siteMetadata.locales.map((listLocale) => (
             <link
               rel="alternate"
               href={`${baseURL}/${
-                locale === siteMetadata.defaultLocale
+                listLocale === siteMetadata.defaultLocale
                   ? '404'
-                  : `${siteMetadata.localePaths[locale.replace('-', '_')]}/404`
+                  : `${siteMetadata.localePaths[listLocale.replace('-', '_')]}/404`
               }`}
-              hrefLang={siteMetadata.simpleLocales[locale.replace('-', '_')]}
-              key={locale}
+              hrefLang={siteMetadata.simpleLocales[listLocale.replace('-', '_')]}
+              key={listLocale}
             />
           ));
         } else {
-          return siteMetadata.locales.map((locale) => (
+          return siteMetadata.locales.map((listLocale) => (
             <link
               rel="alternate"
-              href={createLocaleURL(baseURL, props.pageId, locale, data)}
-              hrefLang={siteMetadata.simpleLocales[locale.replace('-', '_')]}
-              key={locale}
+              href={createLocaleURL(baseURL, pageID, listLocale, data)}
+              hrefLang={siteMetadata.simpleLocales[listLocale.replace('-', '_')]}
+              key={listLocale}
             />
           ));
         }
       })()}
     </Helmet>
   );
-};
+}
+
+Head.propTypes = propTypes;
+Head.defaultProps = defaultProps;
 
 export default Head;