+++ src/components/link/LocalizedLink.jsx
@@ -2,11 +2,12 @@
 // Licensed under the MIT License
 
 import React from 'react';
+import PropTypes from 'prop-types';
 import { Link, useStaticQuery, graphql } from 'gatsby';
 
 import createPagePath from '../../util/createPagePath';
 
-const createPathFromSlug = (slug, locale, data) => {
+const createPathFromSlug = function createPathFromSlugForLocale(slug, locale, data) {
   const { defaultLocale, localePaths } = data.site.siteMetadata;
 
   const authorNodes = data.allContentfulAuthor.edges.filter(({ node }) => node.slug === slug);
@@ -73,7 +74,17 @@ const createPathFromSlug = (slug, locale, data) => {
   return null;
 };
 
-const LocalizedLink = (props) => {
+const propTypes = {
+  children: PropTypes.node.isRequired,
+  className: PropTypes.string,
+  locale: PropTypes.string.isRequired,
+  onClick: PropTypes.func,
+  to: PropTypes.string.isRequired,
+};
+
+const defaultProps = { className: null, onClick: null };
+
+function LocalizedLink({ children, className, locale, onClick, to }) {
   const data = useStaticQuery(
     graphql`
       query {
@@ -103,47 +114,54 @@ const LocalizedLink = (props) => {
 
   const { defaultLocale, localePaths } = data.site.siteMetadata;
 
-  if (props.to === '/') {
+  if (to === '/') {
     return (
       <Link
-        {...props}
-        to={
-          props.locale === defaultLocale ? '/' : `/${localePaths[props.locale.replace('-', '_')]}`
-        }
+        children={children}
+        className={className}
+        onClick={onClick}
+        to={locale === defaultLocale ? '/' : `/${localePaths[locale.replace('-', '_')]}`}
       />
     );
-  } else if (props.to === '/blog') {
-    const blogPath = data.blogPaths.edges.filter(({ node }) => node.node_locale === props.locale)[0]
-      .node;
+  } else if (to === '/blog') {
+    const blogPath = data.blogPaths.edges.filter(({ node }) => node.node_locale === locale)[0].node;
     const pagePath =
-      props.locale === defaultLocale
+      locale === defaultLocale
         ? `/${blogPath.slug}`
-        : `/${localePaths[props.locale.replace('-', '_')]}/${blogPath.slug}`;
-    return <Link {...props} to={pagePath} />;
-  } else if (props.to.startsWith('/')) {
-    const pageSlug = props.to.substring(1);
-    const pagePath = createPathFromSlug(pageSlug, props.locale, data);
+        : `/${localePaths[locale.replace('-', '_')]}/${blogPath.slug}`;
+    return <Link children={children} className={className} onClick={onClick} to={pagePath} />;
+  } else if (to.startsWith('/')) {
+    const pageSlug = to.substring(1);
+    const pagePath = createPathFromSlug(pageSlug, locale, data);
 
     if (pagePath) {
-      return <Link {...props} to={pagePath} />;
+      return <Link children={children} className={className} onClick={onClick} to={pagePath} />;
     } else {
-      return <Link {...props} />;
+      return <Link children={children} className={className} onClick={onClick} />;
     }
+  } else if (to === '404') {
+    const pagePath =
+      locale === defaultLocale ? '/404' : `/${localePaths[locale.replace('-', '_')]}/404`;
+    return <Link children={children} className={className} onClick={onClick} to={pagePath} />;
   } else {
     const node = data.allContentfulEntry.edges.filter(
-      ({ node }) => node.contentful_id === props.to && node.node_locale === props.locale,
+      ({ node }) => node.contentful_id === to && node.node_locale === locale,
     )[0].node;
 
     switch (node.internal.type) {
       case 'ContentfulIndexPage': {
         const indexPath =
-          props.locale === defaultLocale ? '/' : `/${localePaths[props.locale.replace('-', '_')]}`;
-        return <Link {...props} to={indexPath} />;
+          locale === defaultLocale ? '/' : `/${localePaths[locale.replace('-', '_')]}`;
+        return <Link children={children} className={className} onClick={onClick} to={indexPath} />;
       }
-      default:
+      default: {
         break;
+      }
     }
   }
-};
+}
+
+LocalizedLink.propTypes = propTypes;
+LocalizedLink.defaultProps = defaultProps;
 
 export default LocalizedLink;