+++ src/locales/en/headMessages.js
@@ -1,8 +1,9 @@
-// Copyright (c) 2021 Antti Kivi
+// Copyright (c) 2021 Visiosto oy
 // Licensed under the MIT License
 
 export default {
-  metaDescription:
+  headSlogan: 'Entrepreneur from Finland',
+  headDescription:
     'An entrepreneur and doer from Etu-Töölö, Helsinki. I’m all about learning something new every day and smiling as much as possible. In my free time I delight in NGOs and music.',
-  metaOgImageAlt: 'Antti Kivi in a round frame',
+  headOGImageText: 'Antti Kivi in a round frame',
 };