+++ src/locales/fi/footerMessages.js
@@ -3,11 +3,11 @@
 
 export default {
   footerSocialMediaTitle: 'Sosiaalisessa mediassa',
-  footerFacebookAlt: 'Facebookin logo',
-  footerGithubAlt: 'Githubin logo',
-  footerInstagramAlt: 'Instagramin logo',
-  footerLinkedinAlt: 'Linkedinin logo',
-  footerTwitterAlt: 'Twitterin logo',
+  footerFacebookImageText: 'Facebookin logo',
+  footerGithubImageText: 'Githubin logo',
+  footerInstagramImageText: 'Instagramin logo',
+  footerLinkedinImageText: 'Linkedinin logo',
+  footerTwitterImageText: 'Twitterin logo',
   footerOcticons: 'Octicons&nbsp;&copy;&nbsp;2021&nbsp;GitHub,&nbsp;Inc.',
   footerCopyright: '&copy;&nbsp;2021&nbsp;Antti&nbsp;Kivi',
   footerMadeBy: 'Tämä sivusto on tehty <a>Visioston</a> linssin läpi',