+++ src/util/createInternationalization.js
@@ -1,0 +1,6 @@
+// Copyright (c) 2021 Antti Kivi
+// Licensed under the MIT License
+
+export default function createInternationalization(intlUseEffect) {
+  return (k, obj) => intlUseEffect.formatMessage({ id: k }, obj);
+}