+++ src/util/createLocaleURL.js
@@ -3,11 +3,11 @@
 
 // import createPagePath from './createPagePath';
 
-const createLocaleURL = (baseURL, pageId, locale, data) => {
+export default function createLocaleURL(baseURL, pageID, locale, data) {
   const { defaultLocale, localePaths } = data.site.siteMetadata;
 
   const entryNode = data.allContentfulEntry.edges.filter(
-    ({ node }) => node.contentful_id === pageId && node.node_locale === locale,
+    ({ node }) => node.contentful_id === pageID && node.node_locale === locale,
   )[0].node;
 
   switch (entryNode.internal.type) {
@@ -16,9 +16,8 @@ const createLocaleURL = (baseURL, pageId, locale, data) => {
         ? baseURL
         : `${baseURL}/${localePaths[locale.replace('-', '_')]}`;
     }
-    default:
+    default: {
       return null;
+    }
   }
-};
-
-export default createLocaleURL;
+}