+++ src/util/useColorScheme.js
@@ -5,12 +5,13 @@ import { useContext } from 'react';
 
 import ThemeContext from '../components/ThemeContext';
 
-export default () => {
+export default function useColorScheme() {
   if (typeof window !== 'undefined') {
+    // eslint-disable-next-line react-hooks/rules-of-hooks
     const { setColorMode } = useContext(ThemeContext);
 
-    window.matchMedia('(prefers-color-scheme: dark)').addEventListener('change', (e) => {
-      setColorMode(e.matches ? 'dark' : 'light');
+    window.matchMedia('(prefers-color-scheme: dark)').addEventListener('change', (event) => {
+      setColorMode(event.matches ? 'dark' : 'light');
     });
   }
-};
+}